pyproject.toml
[tool.poetry]
name = "lms-project"
version = "0.1.0"
description = ""
authors = ["ildar_safarov <ildarsafarov555@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^6.0.3"
djangorestframework = "^3.17.1"
pillow = "^12.1.1"
django-filter = "^25.2"
djangorestframework-simplejwt = "^5.5.1"
stripe = "^15.0.1"
celery = "^5.6.3"
redis = "^7.4.0"
django-celery-beat = "^2.9.0"
eventlet = "^0.41.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
